# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/check-availability/{key}/{value}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Check for availability",
                "description": "Check for availability of username, email or phone",
                "operationId": "checkAvailability",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "key",
                        "required": true,
                        "example": "email"
                    },
                    {
                        "name": "value",
                        "in": "path",
                        "description": "value",
                        "required": true,
                        "example": "test@kilogram.ga"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "object",
                                            "example": {
                                                "success": true,
                                                "message": "Available to use"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Key can only be phone, email and username",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": {
                                                "success": false,
                                                "message": "Key can only be phone, email and username"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Already in use so not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "object",
                                            "example": {
                                                "success": false,
                                                "message": "Already in use"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}